ctl_zoneinfo.c: squash size_t compiler complaints in map_refresh() and map_free()

+++ imap/ctl_zoneinfo.c
@@ -412,7 +412,7 @@ void do_zonedir(const char *dir, struct hash_table *tzentries,
 	    /* Path is a regular file (zone) */
 	    int fd;
 	    const char *base = NULL;
-	    unsigned long len = 0;
+	    size_t len = 0;
 	    icalcomponent *ical, *comp;
 	    icalproperty *prop;
